requirements.txt
Flask
flask-cors
pytesseract
opencv-python
psycopg2-binary
gunicorn